util: create oc_endpoint_addresses_t
oc_endpoint_addresess_t is generalized oc_cloud_endpoints_t

+++ messaging/coap/coap_internal.h
@@ -105,7 +105,7 @@ typedef struct
   uint8_t token_len;
   uint8_t token[COAP_TOKEN_LEN];
 
-  uint8_t options[COAP_OPTION_SIZE1 / OPTION_MAP_SIZE +
+  uint8_t options[(size_t)COAP_OPTION_SIZE1 / (size_t)OPTION_MAP_SIZE +
                   1]; /* bitmap to check if option is set */
 
   uint16_t content_format; /* parse options once and store; allows setting
